clean up all code with clang format, comments, 60 chars line width

+++ src/rt/rt_math.cpp
@@ -3,7 +3,8 @@
 #include <cmath>
 
 float rt::math::sinusodal(float freqencyHz, float phase) {
-  return sin(freqencyHz * 2 * RT_PI * phase);
+  return static_cast<float>(
+      sin(freqencyHz * 2.f * RT_PI * phase));
 };
 
 bool rt::math::inrange(float a, float lower, float upper) {
